requirements.txt
Django==4.1.5
drf-spectacular==0.27.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-mptt==0.16.0
django-filter==23.5
python-dotenv==0.21.0
psycopg2-binary==2.9.7
django-debug-toolbar==3.8.1
django-cors-headers
text-unidecode
python-barcode==0.15.1
django-celery-beat==2.6.0
django_celery_results
pillow
requests

python-telegram-bot==21.1.1
nest-asyncio==1.6.0
redis
aiohttp